reorder scss file structure and move board components into own folder, rough version of seed-bank screen
