+++ src/scss/_animations.scss
@@ -1,0 +1,98 @@
+@keyframes enter {
+  0% { opacity: 0; }
+  100% { opacity: 1; }
+}
+
+.enter {
+  animation-name: enter;
+  animation-fill-mode: forwards;
+}
+
+@keyframes leave {
+  0% { opacity: 1; }
+  100% { opacity: 0; }
+}
+
+.leave {
+  animation-name: leave;
+  animation-fill-mode: forwards;
+}
+
+@keyframes doubleElasticBounceIn {
+  0% { transform: scale(0) }
+  20% { transform: scale(0.5) }
+  36% { transform: scale(1.5) }
+  52% { transform: scale(0.8) }
+  68% { transform: scale(1.2) }
+  84% { transform: scale(0.9) }
+  100% { transform: scale(1) }
+}
+
+@keyframes elasticBounceIn {
+  0% { transform: scale(0) }
+  50% { transform: scale(1.3) }
+  100% { transform: scale(1) }
+}
+
+@keyframes fall {
+  0% {
+    transform: translateY(0);
+    opacity: 1;
+  }
+  100% {
+    transform: translateY(1000px);
+    opacity: 0;
+  }
+}
+
+// TODO: need to remove, refactor splash screen to use velocity-react
+@for $i from 1 through 12 {
+  $d: $i / 2;
+  .duration-#{$d} {
+    animation-duration: $d + 0.001s;
+  }
+  .delay-#{$d} {
+    animation-delay: $d + 0.001s;
+  }
+}
+
+@keyframes bounceInDown {
+  from, 60%, 75%, 90%, to {
+    -webkit-animation-timing-function: cubic-bezier(0.215, 0.610, 0.355, 1.000);
+    animation-timing-function: cubic-bezier(0.215, 0.610, 0.355, 1.000);
+  }
+  0% {
+    opacity: 0;
+    -webkit-transform: translate3d(0, -3000px, 0);
+    transform: translate3d(0, -3000px, 0);
+  }
+  60% {
+    opacity: 1;
+    -webkit-transform: translate3d(0, 25px, 0);
+    transform: translate3d(0, 25px, 0);
+  }
+  75% {
+    -webkit-transform: translate3d(0, -10px, 0);
+    transform: translate3d(0, -10px, 0);
+  }
+  90% {
+    -webkit-transform: translate3d(0, 5px, 0);
+    transform: translate3d(0, 5px, 0);
+  }
+  100% {
+    -webkit-transform: none;
+    transform: none;
+  }
+}
+
+@for $i from 1 through 7 {
+  @keyframes bounceInDown-#{$i} {
+    $height: ($grid-large / 1.1 - $grid-offset-small) * $i;
+    $offset: 10px;
+
+    0% { transform: translate3d(0, 0, 0) }
+    60% { transform: translate3d(0, ($height + $offset), 0) }
+    85% { transform: translate3d(0, ($height - ($offset / 2) + 2), 0) }
+    100% { transform: translate3d(0, ($height), 0) }
+  }
+}
