--- src/scss/_customAnimations.scss
@@ -1,46 +1,0 @@
-@keyframes enter {
-  0% { opacity: 0; }
-  100% { opacity: 1; }
-}
-
-.enter {
-  animation-name: enter;
-  animation-fill-mode: forwards;
-}
-
-@keyframes leave {
-  0% { opacity: 1; }
-  100% { opacity: 0; }
-}
-
-.leave {
-  animation-name: leave;
-  animation-fill-mode: forwards;
-}
-
-@keyframes doubleElasticBounceIn {
-  0% { transform: scale(0) }
-  20% { transform: scale(0.5) }
-  36% { transform: scale(1.5) }
-  52% { transform: scale(0.8) }
-  68% { transform: scale(1.2) }
-  84% { transform: scale(0.9) }
-  100% { transform: scale(1) }
-}
-
-@keyframes elasticBounceIn {
-  0% { transform: scale(0) }
-  50% { transform: scale(1.3) }
-  100% { transform: scale(1) }
-}
-
-// TODO: need to remove, refactor splash screen to use velocity-react
-@for $i from 1 through 12 {
-  $d: $i / 2;
-  .duration-#{$d} {
-    animation-duration: $d + 0.001s;
-  }
-  .delay-#{$d} {
-    animation-delay: $d + 0.001s;
-  }
-}
