+++ src/scss/board/_weather.scss
@@ -45,17 +45,6 @@
   height: 100vh;
 }
 
-@keyframes fall {
-  0% {
-    transform: translateY(0);
-    opacity: 1;
-  }
-  100% {
-    transform: translateY(1000px);
-    opacity: 0;
-  }
-}
-
 .rain-drop {
   width: 2px;
   height: 20px;
